myrequirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
av==14.3.0
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
datasets==3.5.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.7
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
flash-attn @ file:///home/yuhong_wang/src/flash_attn-2.7.3%2Bcu11torch2.1cxx11abiFALSE-cp311-cp311-linux_x86_64.whl#sha256=ad0aea325c1964d417fc46df76f283863da4158e0a259697796182f100584287
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
hjson==3.1.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
liger_kernel==0.5.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
sentry-sdk==2.27.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tensorboardX==2.6.2.2
tokenizers==0.21.1
torch==2.1.2+cu118
torchaudio==2.1.2+rocm5.6
torchvision==0.16.2+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.1
trl==0.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
